+++ src/main/java/security/packaging/SafetyDbLoadException.kt
@@ -1,0 +1,3 @@
+package security.packaging
+
+class SafetyDbLoadException (message: String): Exception(message)